# swagger_server/models/centre.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.pharmacie_localisation import PharmacieLocalisation  # noqa: F401,E501
from swagger_server import util


class Centre(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: str=None, nom: str=None, tel: int=None, localisation: PharmacieLocalisation=None):  # noqa: E501
        """Centre - a model defined in Swagger

        :param id: The id of this Centre.  # noqa: E501
        :type id: str
        :param nom: The nom of this Centre.  # noqa: E501
        :type nom: str
        :param tel: The tel of this Centre.  # noqa: E501
        :type tel: int
        :param localisation: The localisation of this Centre.  # noqa: E501
        :type localisation: PharmacieLocalisation
        """
        self.swagger_types = {
            'id': str,
            'nom': str,
            'tel': int,
            'localisation': PharmacieLocalisation
        }

        self.attribute_map = {
            'id': 'id',
            'nom': 'nom',
            'tel': 'tel',
            'localisation': 'localisation'
        }
        self._id = id
        self._nom = nom
        self._tel = tel
        self._localisation = localisation

    @classmethod
    def from_dict(cls, dikt) -> 'Centre':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Centre of this Centre.  # noqa: E501
        :rtype: Centre
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Centre.


        :return: The id of this Centre.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Centre.


        :param id: The id of this Centre.
        :type id: str
        """

        self._id = id

    @property
    def nom(self) -> str:
        """Gets the nom of this Centre.


        :return: The nom of this Centre.
        :rtype: str
        """
        return self._nom

    @nom.setter
    def nom(self, nom: str):
        """Sets the nom of this Centre.


        :param nom: The nom of this Centre.
        :type nom: str
        """

        self._nom = nom

    @property
    def tel(self) -> int:
        """Gets the tel of this Centre.


        :return: The tel of this Centre.
        :rtype: int
        """
        return self._tel

    @tel.setter
    def tel(self, tel: int):
        """Sets the tel of this Centre.


        :param tel: The tel of this Centre.
        :type tel: int
        """

        self._tel = tel

    @property
    def localisation(self) -> PharmacieLocalisation:
        """Gets the localisation of this Centre.


        :return: The localisation of this Centre.
        :rtype: PharmacieLocalisation
        """
        return self._localisation

    @localisation.setter
    def localisation(self, localisation: PharmacieLocalisation):
        """Sets the localisation of this Centre.


        :param localisation: The localisation of this Centre.
        :type localisation: PharmacieLocalisation
        """

        self._localisation = localisation